chore: bump dependencies to work with apim 4.0.x BREAKING CHANGE: require java 17, works from apim 4.0.x https://gravitee.atlassian.net/browse/APIM-2218
