--- src/test/resources/apis/api.json
@@ -1,50 +1,0 @@
-{
-  "id": "my-api",
-  "name": "my-api",
-  "gravitee": "2.0.0",
-  "proxy": {
-    "context_path": "/test",
-    "endpoints": [
-      {
-        "name": "default",
-        "target": "http://localhost:8080/team",
-        "http": {
-          "connectTimeout": 3000,
-          "readTimeout": 60000
-        }
-      }
-    ]
-  },
-  "flows": [
-    {
-      "name": "flow-1",
-      "methods": [
-        "GET"
-      ],
-      "enabled": true,
-      "path-operator": {
-        "path": "/",
-        "operator": "STARTS_WITH"
-      },
-      "pre": [
-        {
-          "name": "Compute query param to accept ; as a normal char",
-          "description": "",
-          "enabled": true,
-          "policy": "custom-query-parameters-parser",
-          "configuration": {}
-        }
-      ],
-      "post": [
-        {
-          "name": "Query params to header",
-          "description": "",
-          "enabled": true,
-          "policy": "query-params-to-header",
-          "configuration": {}
-        }
-      ]
-    }
-  ],
-  "resources": []
-}
